Maintenance - Cleanup, Organization & README
- UML Doc(.drawio) for workflow & future plans
- Application: Fullscreen Window & Project loading
- Asset Serializer: AssetPack & Asset::GetByteArray()

+++ editor/assets/shaders/Renderer2D_Sprite.glsl
@@ -6,13 +6,13 @@
 layout(location = 0) in vec3 a_Position;
 layout(location = 1) in vec4 a_Color;
 layout(location = 2) in vec2 a_TextureCoord;
-layout(location = 3) in float a_TextureIndex;
+layout(location = 3) in int a_TextureIndex;
 layout(location = 4) in float a_TilingFactor;
 layout(location = 5) in int a_EntityID;
 
 out vec4 v_Color;
 out vec2 v_TextureCoord;
-flat out float v_TextureIndex;
+flat out int v_TextureIndex;
 out float v_TilingFactor;	
 flat out int v_EntityID;
 
@@ -36,7 +36,7 @@ layout(location = 1) out int entityID;
 
 in vec4 v_Color;
 in vec2 v_TextureCoord;
-flat in float v_TextureIndex;
+flat in int v_TextureIndex;
 in float v_TilingFactor;
 flat in int v_EntityID;
 
@@ -45,7 +45,7 @@ uniform sampler2D u_Textures[32];
 void main()
 {
 	vec4 texColor = v_Color;
-	texColor *= texture(u_Textures[int(v_TextureIndex)], v_TextureCoord * v_TilingFactor);
+	texColor *= texture(u_Textures[v_TextureIndex], v_TextureCoord * v_TilingFactor);
 	
 	if(texColor.a == 0.0)
 	{
